+++ features/booking/entity.go
@@ -12,8 +12,8 @@ type Core struct {
 	CampID        uint
 	Title         string
 	Image         string
-	Latitude      complex64
-	Longitude     complex64
+	Latitude      float64
+	Longitude     float64
 	Address       string
 	City          string
 	CampPrice     int
